+++ packages/typo3-docs-theme/assets/sass/_highlightjs_github.scss
