+++ packages/typo3-docs-theme/assets/sass/codeblock.scss
@@ -1,3 +1,5 @@
+@import '_highlightjs_github';
+
 /* codeblock.css
  *  This file defines syntax highlighting in code blocks.
  *  It has no dependencies and works standalone.
