+++ packages/typo3-docs-theme/resources/template/structure/layout.html.twig
@@ -15,10 +15,10 @@
         {%- if title!=null and env.projectNode.title!=null %} — {% endif -%}
         {%- if env.projectNode.title -%}{{ env.projectNode.title }} {% endif -%} {{ env.projectNode.version }} documentation</title>
 
-    <link href="{{ getRelativePath('_assets/github.css') }}" rel="stylesheet">
-    <link href="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.9.0/css/fontawesome.css" rel="stylesheet">
-    <link href="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.9.0/css/webfonts.css" rel="stylesheet">
-    <link href="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.9.0/css/theme.css" rel="stylesheet">
+    <link href="{{ getRelativePath('_resources/css/fontawesome.css') }}" rel="stylesheet">
+    <link href="{{ getRelativePath('_resources/css/webfonts.css') }}" rel="stylesheet">
+    <link href="{{ getRelativePath('_resources/css/theme.css') }}" rel="stylesheet">
+    <link href="{{ getRelativePath('_resources/css/codeblock.css') }}" rel="stylesheet">
     <link href="https://docs.typo3.org/search/" rel="search" title="Search">
     {% block headerLinks %}
     {% endblock %}
@@ -44,7 +44,7 @@
             <div class="page-header-inner">
                 <!-- pageheader -->
                 <a class="logo" href="https://docs.typo3.org/" title="TYPO3 Documentation">
-                    <img alt="TYPO3 Logo" class="logo-image" src="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.4.3/img/typo3-logo.svg" width="484" height="130">
+                    <img alt="TYPO3 Logo" class="logo-image" src="{{ getRelativePath('_resources/img/typo3-logo.svg') }}" width="484" height="130">
                 </a>
                 <!-- /pageheader -->
             </div>
@@ -164,14 +164,14 @@ j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src=
         HAS_SOURCE: true
     };
 </script>
-<script src="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.9.0/js/jquery.min.js"></script>
-<script src="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.9.0/js/underscore.min.js"></script>
-<script src="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.9.0/js/doctools.min.js"></script>
-<script src="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.9.0/js/popper.min.js"></script>
-<script src="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.9.0/js/bootstrap.min.js"></script>
-<script src="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.9.0/js/theme.min.js"></script>
-<script src="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.9.0/js/searchtools.min.js"></script>
-<script src="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.9.0/js/autocomplete.min.js"></script>
+<script src="{{ getRelativePath('_resources/js/jquery.min.js') }}"></script>
+<script src="{{ getRelativePath('_resources/js/underscore.min.js') }}"></script>
+<script src="{{ getRelativePath('_resources/js/doctools.min.js') }}"></script>
+<script src="{{ getRelativePath('_resources/js/popper.min.js') }}"></script>
+<script src="{{ getRelativePath('_resources/js/bootstrap.min.js') }}"></script>
+<script src="{{ getRelativePath('_resources/js/theme.min.js') }}"></script>
+<script src="{{ getRelativePath('_resources/js/searchtools.min.js') }}"></script>
+<script src="{{ getRelativePath('_resources/js/autocomplete.min.js') }}"></script>
 <script>
     jQuery(function() { Search.loadIndex("searchindex.js"); });
 </script>
