+++ tests/Integration/tests-full/sitemap/expected/index.html
@@ -13,10 +13,10 @@
     <meta content="2023-01-01T12:00:00+00:00" property="article:modified_time">
     <title>Document Title documentation</title>
 
-    <link href="/_assets/github.css" rel="stylesheet">
-    <link href="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.9.0/css/fontawesome.css" rel="stylesheet">
-    <link href="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.9.0/css/webfonts.css" rel="stylesheet">
-    <link href="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.9.0/css/theme.css" rel="stylesheet">
+    <link href="/_resources/css/fontawesome.css" rel="stylesheet">
+    <link href="/_resources/css/webfonts.css" rel="stylesheet">
+    <link href="/_resources/css/theme.css" rel="stylesheet">
+    <link href="/_resources/css/codeblock.css" rel="stylesheet">
     <link href="https://docs.typo3.org/search/" rel="search" title="Search">
                 <link href="/page1.html" rel="next" title="Page 1"/>
             <link href="/index.html" rel="top" title="Document Title"/>
@@ -42,7 +42,7 @@
             <div class="page-header-inner">
                 <!-- pageheader -->
                 <a class="logo" href="https://docs.typo3.org/" title="TYPO3 Documentation">
-                    <img alt="TYPO3 Logo" class="logo-image" src="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.4.3/img/typo3-logo.svg" width="484" height="130">
+                    <img alt="TYPO3 Logo" class="logo-image" src="/_resources/img/typo3-logo.svg" width="484" height="130">
                 </a>
                 <!-- /pageheader -->
             </div>
@@ -210,14 +210,14 @@ j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src=
         HAS_SOURCE: true
     };
 </script>
-<script src="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.9.0/js/jquery.min.js"></script>
-<script src="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.9.0/js/underscore.min.js"></script>
-<script src="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.9.0/js/doctools.min.js"></script>
-<script src="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.9.0/js/popper.min.js"></script>
-<script src="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.9.0/js/bootstrap.min.js"></script>
-<script src="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.9.0/js/theme.min.js"></script>
-<script src="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.9.0/js/searchtools.min.js"></script>
-<script src="https://typo3.azureedge.net/typo3documentation/theme/sphinx_typo3_theme/4.9.0/js/autocomplete.min.js"></script>
+<script src="/_resources/js/jquery.min.js"></script>
+<script src="/_resources/js/underscore.min.js"></script>
+<script src="/_resources/js/doctools.min.js"></script>
+<script src="/_resources/js/popper.min.js"></script>
+<script src="/_resources/js/bootstrap.min.js"></script>
+<script src="/_resources/js/theme.min.js"></script>
+<script src="/_resources/js/searchtools.min.js"></script>
+<script src="/_resources/js/autocomplete.min.js"></script>
 <script>
     jQuery(function() { Search.loadIndex("searchindex.js"); });
 </script>
